Correct types via cast for windows; strlen is type size_t in windows, not int.
git-svn-id: file:///home/jj/hercules.svn/trunk@6007 956126f8-22a0-4046-8f4a-272fa8102e63

+++ ecpsvm.c
@@ -2877,7 +2877,7 @@ ECPSVM_CMDENT *ecpsvm_getcmdent(char *cmd)
         ce=&ecpsvm_cmdtab[i];
         if(strlen(cmd)<=strlen(ce->name) && strlen(cmd)>=(size_t)ce->abbrev)
         {
-            clen=strlen(cmd);
+            clen=(int)strlen(cmd);
             if(strncasecmp(cmd,ce->name,clen)==0)
             {
                 return(ce);
